+++ Sources/Views/ALKLinkPreviewBaseCell.swift
@@ -15,6 +15,7 @@ class ALKLinkPreviewBaseCell: ALKMessageCell {
         super.update(viewModel: viewModel, messageStyle: messageStyle, mentionStyle: mentionStyle)
         linkView.setLocalizedStringFileName(localizedStringFileName)
         url = ALKLinkPreviewManager.extractURLAndAddInCache(from: viewModel.message, identifier: viewModel.identifier)?.absoluteString
+        linkView.previewImageView.image = UIImage(named: "default_image", in: Bundle.km, compatibleWith: nil)
     }
 
     override func setupViews() {