+++ CHANGELOG.md
@@ -4,6 +4,10 @@ The changelog for [KommunicateChatUI-iOS-SDK](https://github.com/Kommunicate-io/
 
 ## Unreleased
 - Zendesk (Zopim) integration optimisations
+## [1.2.6] 2024-01-05
+- Fixed link preview showing for deep link
+- Added the caption Screen for Attachment.
+- Fixed New Form UI Cutting from Bottom.
 
 ## [1.2.5] 2023-12-23
 - Restrict agent reply for zendesk conversation